Remove ZcashWallet in favor of self-contained ZcashAccount
